requirements.txt
# Compatibility Pipeline - Python Dependencies
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Configuration and serialization
pyyaml>=5.4
joblib>=1.1.0
